+++ game/utility/generator/CMakeLists.txt
@@ -1,0 +1,11 @@
+cmake_minimum_required (VERSION 3.15)
+
+project ("generator")
+
+add_library (${PROJECT_NAME} INTERFACE)
+
+target_include_directories(${PROJECT_NAME}
+    INTERFACE
+        $<INSTALL_INTERFACE:include>    
+        $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+	)
