+++ game/utility/tostring/include/ToString.h
@@ -1,0 +1,14 @@
+#pragma once
+#include <sstream>
+
+
+
+// Since std::to_string doesn't work on MinGW we have to implement
+// our own to support all platforms.
+template <typename T>
+std::string toString(const T& value)
+{
+	std::stringstream stream;
+	stream << value;
+	return stream.str();
+}
