+++ game/view/label/include/Label.h
@@ -1,0 +1,23 @@
+#pragma once
+#include <SFML/Graphics/Font.hpp>
+#include <SFML/Graphics/Text.hpp>
+#include <string>
+#include "Widget.h"
+#include "ToString.h"
+
+
+
+class Label : public Widget
+{
+public:
+    Label (sf::Vector2f);
+
+    virtual void draw (sf::RenderTarget&, sf::RenderStates) const final;
+
+    void            setText (std::string);
+    virtual void    setPosition (sf::Vector2f) final;
+
+private:
+    sf::Font font_;
+    sf::Text text_;
+};
