--- src/Graphics/Widget.cpp
@@ -1,11 +1,0 @@
-#include "Graphics/Widget.hpp"
-
-Widget::Widget (sf::Vector2f position) {
-    position_ = position;
-}
-
-Widget::~Widget () {}
-
-sf::Vector2f Widget::getPosition () const {
-    return position_;
-}
